pyproject.toml
[tool.poetry]
name = "red-hat"
version = "0.1.0"
description = ""
authors = ["edg956 <doniakeu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
requests = "^2.27.1"
ratelimit = "^2.2.1"
tenacity = "^8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
